eth-market-forecasting/backend/requirements.txt
# Auto-generated file
pandas
requests
dash
flask
python-dotenv
scikit-learn
sqlite3
requests==2.28.1
pandas==1.5.3
scikit-learn==1.1.3
numpy==1.23.5
dash==2.9.0
dash-core-components==2.0.0
dash-html-components==2.0.0
flask==2.2.2
xgboost 
lightgbm
joblib